requirements.txt
flask==3.1.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.11
gunicorn
SQLAlchemy==2.0.29
alembic==1.13.1
flasgger==0.9.7.1
sentry-sdk[flask]==2.27.0
pytest-mock